(MAINT) Update CI workflow input
Prior to this commit acceptance tests were not triggering when the value
of `with_acceptance` was set to `true`. This is true for the default
value and user provided values.

It turns out that this is because of context differences between events.
Namely `workflow_call` and `workflow_dispatch`.

With `workflow_dispatch` events, inputs come from
`github.event.inputs.<input_name>' and boolean values are treated as
strings.

With `workflow_call` events, inputs come from `inputs.<input_name>`
and boolean values are treated as booleans.

I would still be stuck in a hole if it wasn't for this blog:

https://medium.com/@sohail.ra5/github-actions-passing-boolean-input-variables-to-reusable-workflow-call-42d39bf7342e