(maint) - Update gem_release_prep.yml
Prior to this commit, the gem release prep workflow required the
verion.rb file to be in ./lib/$gem_name/version.rb. This was determined
to not be suitable, as not all gems followed this folder structure.

This commit introduces a better method, that will work for all of the
gems that will use this workflow. As long as the version.rb file
lives as a file or within a subdirectory of the './lib/' directory,
the workflow will update the version.rb file.