(feat) - add matrix as input
This commit allows a custom matrix to be passed to
module_acceptance.yml.

This allows the calling workflow to pass a matrix that differs from the
one that would be generated by `bundle exec matrix_from_metadata_v2`,
which may be useful when requiring a different provisioner, image type
etc.