(CAT-1618) - Fix issue with code coverage in nightlies
Prior to this commit, code coverage reports would attempt to upload from
the nightlies despite no code coverage reports being generated. This is
because we only run code coverage in the ci.yml job, and not
nightly.yml.

This PR adds another conditional to the upload coverage reports to
Codecov step, to ensure this step is only run when a coverage rake_task
has been supplied.