(CAT-1829) - Update to dynamic checkout ref gem_ci
This commit implements a dynamic checkout ref for the gem_ci workflow.
This is because we will be updating the calling workflows to be run on
the pull_request_target trigger so they can safely access secrets, so
this extra checkout logic is needed in order to checkout the head
branch.